+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### Updated
+
+## [1.4.0]
+
+-   Changed actionbar glow to be more visible.
+-   Updated version checks
+-   Fix #35 - Correctly move the new totem tracker for shaman and DK's
+
 ## [1.3.9]
 
 ### Updated